requirements.txt
streamlit==1.19.0
tensorflow==2.12.0
numpy==1.24.3
opencv-python==4.8.0.76
Pillow==10.0.0